Added EquelleRuntimeCPU class and related types.
The EquelleRuntimeCPU class and the types and classes defined with it in
its header are intended to make it simple to write the C++ code corresponding
to Equelle code. Equelle built-ins are directly represented in the class,
with a C++ interface that is close to Equelle.

For future back-ends, we can consider reusing the interface provided here.