Added random access to Vector types.
Notation is foo[INT] if foo is a Collection of Vector. The integer
can be 0, 1 or 2, and a requirement mechanism has been put in to ensure
that you cannot run a simulator with a grid that has too few dimensions
(if you ever use foo[2] for example, the grid must have 3 dimensions).

The codegen parts (and back-end) must change when we add a way to create Vectors
(and not just Collections Of them), for now an Eigen::Array::col() call is emitted.