First commit of the Equelle kitchen-sink web demonstration
The webdemo folder now contains a vagrant virtual machine setup which runs a webserver hosting the Equelle kitchen-sink demonstration.
The purpose of this demonstration is to provide an environment for testing the Equelle DSL in a browser, without the need to install or compile anything.
The user writes valid Equelle code in the browser code editor and supplies inputs to the simulation, the simulator is then compiled and run on the webserver and outputs the simulation results in the browser for the user to view.